Website Logout working
Logout and login working but using back arrow from index has a half a second view error
